Allow multiple content types (breaking change)
This is a potential breaking change that allows users to include multiple
content types. The `params.contentTypeName` should now be a list of strings,
with each string being the desired content type to be displayed.

This change supports both the home page listing (which will list recent items
from all content types) and individual pages (which will list recent items for
only that specific content type).